Got retasking ships working. Decided to switch to a collectionview as it just make the whole process simpler and allowed me to stick to the original design document, which keeps the design of the app consistent throughout.
Retasking ships between groups is now working. I opted to have the navigation go back only one layer (to the ship detail of the ship that has just been retasked) rather than two (to the task group that the ship has just been retasked out of). This is to reduce potential frustration over accidentally retasking a ship to the wrong group, which if navigation took the user back two layers would require them to navigate back even further and then navigate in 3 layers to correct.

Also I know it's dumb to put this explaination in a commit message rather than some other form of note/comment. I am strapped for time and this is what is convienent.